clip/requirements.txt
numpy>=2.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
transformers>=4.35.0
pandas>=2.1.0
tqdm>=4.66.0
